weird coercion issue

+++ util/token/token.go
@@ -114,13 +114,13 @@ func FromString(accessToken string, secret []byte) (*Token, error) {
 		return nil, err
 	}
 
-	jwt.Claims["bit"] = jwt.Claims["bit"].(int64)
+	jwt.Claims["bit"] = int64(jwt.Claims["bit"].(float64))
 
 	// tok.Name = jwt.Claims["name"].(string)
 	// tok.IssuedAt = time.Unix(int64(jwt.Claims["iat"].(float64)), 0)
 	tok.Id = jwt.Claims["jti"].(string)
 	tok.EntityId = jwt.Claims["sub"].(string)
-	tok.Permissions = bit.Field(jwt.Claims["bit"].(float64))
+	tok.Permissions = bit.Field(jwt.Claims["bit"].(int64))
 	tok.jwt = jwt
 	tok.Secret = secret
